pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "c7n-azure-container-apps"
version = "0.1.0"
description = "Cloud Custodian Azure implementation for Azure Container Apps Jobs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.14"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["cloud-custodian", "azure", "container-apps", "policy", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "c7n>=0.9.47",
    "c7n-azure>=0.7.46",
    "click>=8.1.0",
    "azure-identity>=1.15.0",
    "azure-storage-queue>=12.9.0",
    "azure-storage-blob>=12.19.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
c7n-azure-runner = "c7n_azure_container_apps.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/c7n-azure-container-apps"
Documentation = "https://github.com/your-org/c7n-azure-container-apps#readme"
Repository = "https://github.com/your-org/c7n-azure-container-apps"

[tool.hatch.build.targets.wheel]
packages = ["src/c7n_azure_runner", "src/c7n_azure_container_apps"]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["c7n_azure_runner", "c7n_azure_container_apps"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["c7n_azure_runner", "c7n_azure_container_apps"]
mypy_path = "src"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
